Create LICENSE.md
Added rider that LexDAO curates the metadata which is hereby released to the public domain but makes *NO* assertion as to the original content (which may be copyrighted eg ISO standards documents), and is provided on a good-faith basis. Members may charge extra for reasonable commercial effort to review and update as to the changing legal landscape.